Moved queries --> graph_client
