--- lib/client.dart
@@ -1,18 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:graphql_flutter/graphql_flutter.dart';
-
-class Config {
-  static final link = HttpLink(
-    uri: 'http://10.0.0.213:4000/',
-  );
-
-  static ValueNotifier<GraphQLClient> initializeClient() {
-    ValueNotifier<GraphQLClient> client = ValueNotifier(
-      GraphQLClient(
-        cache: InMemoryCache(),
-        link: link,
-      ),
-    );
-    return client;
-  }
-}
